ID-1: created folders account, misc, pages and poetry and moved pages accordingly
