+++ src/account/ProfilePage.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import StaticContentService from './StaticContentServiceYaml'
+import StaticContentService from '../StaticContentServiceYaml'
 
 // for formatting
-import './pages/TopSectionMainPage/TopSectionMainPage.css';
+import './../pages/TopSectionMainPage/TopSectionMainPage.css';
 
-import PoemPage from './PoemPage';
+import PoemPage from '../poetry/PoemPage';
 
 import ReactTable from 'react-table'
 import 'react-table/react-table.css'
