+++ src/misc/CommentPage.js
@@ -1,9 +1,9 @@
 import React from 'react'
-import StaticContentService from './StaticContentServiceYaml'
+import StaticContentService from '../StaticContentServiceYaml'
 // import Tabs from './Tabs'
 
 // for formatting
-import './pages/TopSectionMainPage/TopSectionMainPage.css';
+import './../pages/TopSectionMainPage/TopSectionMainPage.css';
 
 import Tab from 'react-bootstrap/Tab'
 import Tabs from 'react-bootstrap/Tabs'
@@ -192,7 +192,7 @@ class CommentsPage extends React.Component {
                 success: (data) => { // success funciton starts
                     var sherArray = sherName.split("_");
 
-                    const yamlFile = require('!raw-loader!./assets/poems/' + sherArray[0] + '/' + sherArray[0] + '_' + sherArray[1] + '.yaml');
+                    const yamlFile = require('!raw-loader!./../assets/poems/' + sherArray[0] + '/' + sherArray[0] + '_' + sherArray[1] + '.yaml');
 
                     console.log("After calling yamlFiles");
                     console.log("Value of yamlFile");
