+++ src/poetry/ListPoemPage.js
@@ -1,8 +1,8 @@
 import React from 'react'
-import StaticContentService from './StaticContentServiceYaml'
+import StaticContentService from '../StaticContentServiceYaml'
 
 // for formatting
-import './pages/TopSectionMainPage/TopSectionMainPage.css';
+import './../pages/TopSectionMainPage/TopSectionMainPage.css';
 
 import Divider from '@material-ui/core/Divider';
